Add files via upload
I don't know how to add another project of the solution to source control, that's why I am just aploading it as a ZIP file.
You can download it and unpack(!) it in a folder of your choice.
Then you open your ZPool solution in Visual Sutio, right click on the Solution name
Solution 'ZPoolSolution' > Add > Existing Project (see attached image: existingproject.png)
Find the folder with the unpacked unit test project, select the .csproj file and add it.
Should work. Or might need to add a reference to the main project.